requirements.txt
pytest
playwright
requests
python-dotenv
allure-pytest